truthbrush/requirements.txt
truthbrush
python-dotenv
tenacity
flask>=2.3.0
plotly>=5.0.0
pandas>=1.5.0
networkx>=3.0